Tests/LibWeb/Text/expected/DOM/Event-has-unforgeable-properties.txt

isTrusted exists on Event instance itself: true
isTrusted does not exist on LocationPrototype: true
isTrusted descriptor is not undefined: true
isTrusted enumerable: true
isTrusted configurable: false
isTrusted writable: undefined
isTrusted exists on UIEvent instance itself: true
isTrusted does not exist on LocationPrototype: true
isTrusted descriptor is not undefined: true
isTrusted enumerable: true
isTrusted configurable: false
isTrusted writable: undefined
isTrusted exists on CustomEvent instance itself: true
isTrusted does not exist on LocationPrototype: true
isTrusted descriptor is not undefined: true
isTrusted enumerable: true
isTrusted configurable: false
isTrusted writable: undefined
